Added purchase service tests
