--- backend/src/test/java/com/revature/services/OrderServiceTest.java
@@ -1,37 +1,0 @@
-package com.revature.services;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class OrderServiceTest {
-    @Test
-    void testDelete() {
-
-    }
-
-    @Test
-    void testFindAll() {
-
-    }
-
-    @Test
-    void testFindAllByUserId() {
-
-    }
-
-    @Test
-    void testFindById() {
-
-    }
-
-    @Test
-    void testSave() {
-
-    }
-
-    @Test
-    void testSaveAll() {
-
-    }
-}
