+++ backend/src/test/java/com/revature/services/OrderServiceTests.java
@@ -4,7 +4,27 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class PurchaseServiceTest {
+public class OrderServiceTests {
+    @Test
+    void testDelete() {
+
+    }
+
+    @Test
+    void testFindAll() {
+
+    }
+
+    @Test
+    void testFindAllByUserId() {
+
+    }
+
+    @Test
+    void testFindById() {
+
+    }
+
     @Test
     void testSave() {
